Add the ability to display admoninations so they stand out.
This makes a slight change to the pom config, and surrounds the
previous items marked as NOTICE with the note tag.  When
generated now, the Admonination is display with a graphic
icon to help make it stand out from the document.

Same thing can be done for tips, warnings, and caution tags.